feat(client/slider) add slider
