+++ client/src/styles/components/_slider.scss
@@ -4,6 +4,7 @@
     height: 500px; /* Puedes ajustar la altura deseada */
     overflow: hidden;
     margin: 0px 0px 30px 0px;
+    box-shadow: inset 0 0 10px rgba(0, 0, 0, 0.5);
   }
   
   .slider__image {
@@ -24,12 +25,13 @@
     background-color: transparent;
     cursor: pointer;
     z-index: 10;
-    color: #79be21;
+    color: #000;
+    opacity: 0.5;
   }
 
   .slider__prev-button:hover,
   .slider__next-button:hover{
-    color: #fff;
+    color: #00BF63;
   }
   
   .slider__prev-button {
@@ -54,7 +56,7 @@
     width: 10px;
     height: 10px;
     border-radius: 50%;
-    background-color: #fff;
+    background-color: #000000;
     margin: 0 4px;
     cursor: pointer;
     opacity: 0.6;
